refactor pomo to use countdown threads and events
